pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "beat-grid-confidence"
version = "0.1.0"
description = "Confidence-aware beat grid prediction extending beat_this with region-level ambiguity estimation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Rowdi Labs", email = "research@rowdilabs.com" },
]
keywords = ["beat-tracking", "music-information-retrieval", "confidence-estimation", "beat-grid"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0",
    "torchaudio>=2.0",
    "pytorch-lightning>=2.0",
    "beat_this @ git+https://github.com/CPJKU/beat_this.git",
    "mir_eval>=0.7",
    "numpy>=1.24",
    "scipy>=1.10",
    "librosa>=0.10",
    "onnx>=1.14",
    "onnxruntime>=1.16",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
    "mypy>=1.0",
]
experiment = [
    "wandb>=0.16",
    "matplotlib>=3.7",
    "seaborn>=0.13",
    "jupyter>=1.0",
]

[project.urls]
Repository = "https://github.com/Rowdi-Labs/beat-grid-confidence"
Issues = "https://github.com/Rowdi-Labs/beat-grid-confidence/issues"

[project.scripts]
bgc-train = "scripts.train:main"
bgc-evaluate = "scripts.evaluate:main"
bgc-export = "scripts.export_onnx:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
